requirements.txt
# Web Framework
streamlit==1.29.0

# Image Processing
Pillow==10.2.0

# Data Science and Machine Learning
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2

# Deep Learning
tensorflow==2.12.0
keras==2.12.0
torch==2.0.1

# Natural Language Processing
nltk==3.8.1
spacy==3.5.2

# Web Development and Utilities
requests==2.31.0
plotly==5.15.0

# Development and Debugging
pytest==7.3.1
jupyter==1.0.0 